move Metrics class to core
