+++ src/extraction/evaluation/layer_evaluator.py
@@ -7,8 +7,8 @@ from typing import Any
 
 import Levenshtein
 
+from core.benchmark_utils import Metrics
 from extraction.evaluation.benchmark.metrics import OverallMetrics
-from extraction.evaluation.evaluation_dataclasses import Metrics
 from extraction.features.predictions.borehole_predictions import (
     BoreholePredictionsWithGroundTruth,
     FileLayersWithGroundTruth,
